make a table component
