+++ src/utils/common/constants.ts
@@ -10,3 +10,5 @@ export const standardFee = "5000"; // may need adjustment for different tx types
 export const standardDeadline = 1440; // deadline for confirmation, required in most (all?) txs
 export const userLocale = { localeStr: "en-US", options: { timeZone: "America/Chicago" } }; // CST for testing, controls date/time stamps for human readability
 export const unitPrecision = 8; // 8 digits of unit precision, used broadly
+
+export const TableRowsPerPageOptions = [3, 5];
